refactor: rename test folder to run
